requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.46
databases==0.5.5
asyncpg==0.27.0
torch==2.0.0+cpu          # Use the appropriate CUDA version if needed
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.1.3
pyyaml==6.0
matplotlib==3.6.2
jupyterlab==3.5.0
